requirements.txt
groundingdino-py==0.4.0
opencv-python~=4.11.0.86
torch~=2.7.0
torchvision~=0.22.0
requests~=2.32.3
numpy~=2.0.2
pillow~=11.2.1
huggingface-hub~=0.31.2
tqdm~=4.67.1
supervision~=0.25.1
setuptools~=78.1.0
PyYAML~=6.0.2
transformers~=4.51.3
addict~=2.4.0
yapf~=0.43.0
matplotlib~=3.9.4
pycocotools~=2.0.8
timm~=1.0.15
ultralytics~=8.3.135
segment-anything~=1.0